backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
pydantic==2.9.2
pydantic-settings==2.6.0
anthropic>=0.97.0
voyageai==0.3.2
python-frontmatter==1.1.0
markdown-it-py==3.0.0
celery==5.4.0
redis==5.1.1
httpx==0.27.2
authlib==1.3.2
PyJWT==2.9.0
python-multipart==0.0.12
sentence-transformers==3.2.1
diff-match-patch==20241021
